streams/src/requirements.txt
Flask[async]==2.2.3
#Flask-uWSGI-WebSocket==0.6.1 
Websockets==10.4
google-cloud-speech==2.19.0
